requirements.txt
pennylane==0.33.1
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.4
numpy==1.24.4
fastapi==0.104.1
uvicorn==0.24.0
joblib==1.3.2
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6
pytest==7.4.3